pyproject.toml
[build-system]
requires = ["setuptools>=69.5.1"]
build-backend = "setuptools.build_meta"

[project]
name = "sage"
version = "0.1.0"
description = "Sage - an all purpose open-source Discord bot working under the AGPLv3 license."
authors = [{ name = "Rin Nascimento", email = "rin@kmrin.net" }]
readme = "README.md"
requires-python = ">=3.11"
license = "AGPL-3.0-or-later"
dependencies = [
    "alembic>=1.17.2",
    "asyncpg>=0.30.0",
    "discord-py>=2.6.4",
    "pydantic>=2.12.4",
    "python-dotenv>=1.2.1",
    "pyyaml>=6.0.3",
    "sqlalchemy>=2.0.44"
]

[project.optional-dependencies]
dev = [
    "faker>=38.2.0"
]

[project.scripts]
sage = "sage.client:main"

[project.urls]
Source = "https://github.com/kmrin/sage"
TermsOfService = "https://github.com/kmrin/sage/blob/main/sage/assets/docs/terms_of_service.md"
PrivacyPolicy = "https://github.com/kmrin/sage/blob/main/sage/assets/docs/privacy_policy.md"